Change Integer to Long in some classes

+++ src/main/java/io/voucherify/client/model/customer/response/CustomerCampaignLoyalty.java
@@ -13,8 +13,8 @@ import lombok.ToString;
 @ToString
 public class CustomerCampaignLoyalty {
 
-  private Integer points;
+  private Long points;
 
   @JsonProperty("referred_customers")
-  private Integer referredCustomers;
+  private Long referredCustomers;
 }
